% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/table_output.R
\name{Escribano_1}
\alias{Escribano_1}
\title{Escribano model 1}
\usage{

  Escribano_1(path = "C:/Users/akl/Dropbox/Economics/Final_Thesis/Thesis/Tables",
  country = "de")
}
\description{
Escribano model 1
}

