% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/table_output.R
\name{Escribano_2}
\alias{Escribano_2}
\title{Escribano model 2}
\usage{

  Escribano_2(path = "C:/Users/akl/Dropbox/Economics/Final_Thesis/Thesis/Tables",
  country = "de")
}
\description{
Escribano model 2
}

