% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/table_output.R
\name{Escribano_3}
\alias{Escribano_3}
\title{Escribano model 3}
\usage{

  Escribano_3(path = "C:/Users/akl/Dropbox/Economics/Final_Thesis/Thesis/Tables",
  country = "de")
}
\description{
Escribano model 3
}

