% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ewma}
\alias{ewma}
\title{Function for calculating exponentially weighted moving average. Written in
C++.}
\usage{
ewma(x, lambda = 0.975)
}
\arguments{
\item{x}{An atomic vector}

\item{lamb}{A double}
}
\description{
Function for calculating exponentially weighted moving average. Written in
C++.
}

