% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/likelihood functions.R
\name{log_lik_model_a}
\alias{log_lik_model_a}
\title{Model estimation for Escribano2012 model a}
\usage{
log_lik_model_a(input_frame)
}
\arguments{
\item{theta}{An atomic vector of coefficients}

\item{price}{An atomic vector with the price series}
}
\description{
Model estimation for Escribano2012 model a
}

