% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/likelihood_functions.R
\name{log_lik_model_c}
\alias{log_lik_model_c}
\title{Model estimation for Escribano2012 model c}
\usage{
log_lik_model_c(input_frame_s, input_frame_i,
  path = "C:/Users/akl/Dropbox/Economics/Final_Thesis/Thesis/Tables",
  country = "de")
}
\description{
Model estimation for Escribano2012 model c
}

