% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loglik_a}
\alias{loglik_a}
\title{Loglikelihood function for Escribano2012 model a}
\usage{
loglik_a(theta, price, week_dum)
}
\arguments{
\item{theta}{A parameter vector}

\item{price}{A price vector}

\item{week_dum}{A week dummy matrix}
}
\description{
Loglikelihood function for Escribano2012 model a
}

