% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_neg}
\alias{plot_neg}
\title{Function plotting negative prices for dayahed data.}
\usage{
plot_neg(input_frame_spot, file_name, save_path, do_print = FALSE)
}
\arguments{
\item{file_name}{A string indicating name of output}

\item{save_path}{A string with path to save file}

\item{do_print}{A Bolean indicating if plot should be printed}

\item{input_frame}{A dataframe with a series named price}
}
\description{
Function plotting negative prices for dayahed data.
}

