% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_spike}
\alias{plot_spike}
\title{Function plotting spikes of intraday and dayahed data for a given
interval.}
\usage{
plot_spike(input_frame_spot, input_frame_intraday, file_name, save_path,
  do_print = FALSE)
}
\arguments{
\item{file_name}{A string indicating name of output}

\item{save_path}{A string with path to save file}

\item{do_print}{A Bolean indicating if plot should be printed}

\item{input_frame}{A dataframe with a series named price}
}
\description{
Function plotting spikes of intraday and dayahed data for a given
interval.
}

