% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_outputs.R
\name{power_characteristics}
\alias{power_characteristics}
\title{Function generating the plots used in the seasonality subsection of thesis}
\usage{
power_characteristics(input_frame_spot, input_frame_intraday,
  save_path = "C:/Users/akl/Dropbox/Economics/Final_Thesis/Thesis/Child Documents/A_Brief_Overview_of_Electricity_Markets/Figures",
  do_print = FALSE)
}
\arguments{
\item{input_frame_spot}{A data frame with spot price data.}

\item{input_frame_intraday}{A data frame with spot price data.}

\item{save_path}{A string with the path to save plot.}

\item{do_print}{A boolean of whether or not to print the output.}
}
\description{
Function generating the plots used in the seasonality subsection of thesis
}

