% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/forecasting_output.R
\name{read_and_munge_forecast_hour_4}
\alias{read_and_munge_forecast_hour_4}
\title{Read and munge forecasting data}
\usage{
read_and_munge_forecast_hour_4()
}
\description{
Read and munge forecasting data
}

