% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scraping.R
\name{scrape_trans_eex}
\alias{scrape_trans_eex}
\title{Scrape EEX transparancy}
\usage{
scrape_trans_eex(date_from = "2010-01-01", date_to = "2011-07-07")
}
\arguments{
\item{date_from}{A date string of format "yyyy-mm-dd"}

\item{date_to}{A date string of format "yyyy-mm-dd"}
}
\description{
Scrape EEX transparancy
}

